requirements.txt
streamlit
supabase
pandas
streamlit-geolocation
folium
streamlit-folium
opencv-python-headless
zxing-cpp
Pillow